Updated to-do list. Removed KTX2 CMake option temporarily. Doesn't do anything anyways. Added polymorphic allocator class/interface. Removed several helper functions related to enums (i.e toString), will re add if a has been found. Removed the COUNT identifier from public enums. Renamed LoadResult<T> to ResultValue<T> Added the LoadSettings struct. It's currently empty, doesn't do anything. Removed ResultType::NoIdea. Updated several comments describing structs and functions. Added the Texas::Texture abstraction class. Removed calcSingleImageDataSize(MetaData). It's of little use since it doesn't take mip into account. Corrected type of MemReqs_PNG_BackendData::plteChunkDataLength from u64 to u32. Readded the isCompressed(PixelFormat) function. Fixed a performance bug in KTX loading backend related to some debugging. Removed the need for std::abs in the PNG backend. Improved an error message in the PNG backend. Moved functions out namespace Texas brackets in Texas.cpp. Removed the Tools namespace.

+++ include/Texas/detail/MemReqs_BackendData.hpp
@@ -18,7 +18,7 @@ namespace Texas::detail
         const unsigned char* idatChunkStart = nullptr;
         std::uint32_t idatChunkCount = 0;
         const unsigned char* plteChunkStart = nullptr;
-        std::uint64_t plteChunkDataLength = 0;
+        std::uint32_t plteChunkDataLength = 0;
     };
 
     struct MemReqs_BackendData
